Adicao do datatables; refatoracao do carrinho de compras, divisao do seeders

+++ database/migrations/2017_06_20_134151_create_order_product.php
@@ -15,7 +15,7 @@ class CreateOrderProduct extends Migration
     {
         Schema::create('order_product', function (Blueprint $table) {
             $table->increments('id');
-            $table->double('total');
+            $table->decimal('total', 10, 2);
             $table->integer('qtd_itens');
             $table->integer('order_id')->unsigned();
             $table->integer('product_id')->unsigned();
